Change node uuid & name to be based on file id & vm.uuid (root)
The file ID is the same for all files of a snapshot.
Therefore, adding or deleting one of these files does not affect
the snapshot UUID or name. Root files may not have an ID; inn this case,
the VM UUID is used to generate the snapshot UUID and name.